Separate removed inheritance from hidden
The `removed` property is used in a number of places in Metalava but
mainly in `ApiPredicate`. It traverses up the class nesting hierarchy
OR-ing the value at each level to get the actual value it uses. That
is slightly different behavior to that implemented in
`ApiVariantSelectors` which will not inherit it if the `Item` has a
show or hide annotation.

The `ApiPredicate` behavior is what wins out and it also makes the most
sense.

So, this change makes the `ApiVariantSelectors` behave like
`ApiPredicate` by handling it completely separately to the other
properties in `inheritInto()`, moving the logic into its
`initialValueProvider`. That dedups the inheritance logic, makes it
simpler and cleaner and avoids having to update the other properties.

There is the possibility that this will change the behavior of other
parts of Metalava which access `removed` directly instead of via
`ApiPredicate`. However, it will only make them more accurately reflect
the actual `removed` status that gets used when writing out the
signature files.

Bug: 351118278
Test: ./gradlew
Change-Id: Iebe91407f4a6a16632a353ef960d0ea79fa18aa3