Simplify removed handling in ApiPredicate
There is no need to traverse up the hierarchy to find if any
containing `Item` is marked as `@doconly` as that work has already been
done. This change removes that logic and returns `false` immediately if
`removed` is not being ignored and `Item`s `removed` status does not
match was is required. It returns after other early exits to avoid
changing behavior as some of them can return `true`.

Bug: 351118278
Test: ./gradlew
Change-Id: I806dcc0fd229f7b2312354445caae36195e40045