Pull defaultValue: DefaultValue up into ParameterItem
Previously, it was an implementation detail of `DefaultParameterItem`
but it needs to be exposed in `ParameterItem` so that a snapshot of it
can be taken. This change pulls it up and creates an implementation of
the `defaultValue` property for Psi that is based on the methods on the
`PsiParameterItem`. A future change will add a `PsiDefaultValue`..

In order to differentiate between `val defaultValue: DefaultValue` and
`fun defaultValue(): String?` the latter was renamed to
`defaultValueAsString` along with the property in `PsiParameterItem`.

Bug: 353737744
Test: ./gradlew
Change-Id: I414f81053cb8f30a30325530d5de2aeabb672240