Add CodebaseAssembler/Factory abstraction to simplify Codebase creation
Previously, `TurbineSourceParser` would create a `TurbineBasedCodebase`
directly and then call `initialize(...)` on that which would then
create a `TurbineCodebaseInitializer` and call `initialize(...)` on
that which would then store the `codebase` in a property and then bind
the classes.

This change adds a `CodebaseAssembler` and `CodebaseAssemblerFactory`
to allow the `TurbineCodebaseInitializer` to be created during the
`TurbineBasedCodebase` initialization removing the need for the
`TurbineBasedCodebase.initialize()` method.

Although, this is just being used for Turbine here, the intention is
that this pattern will be used by other models as well, to ensure
consistency by enabling improved sharing of code across models. This is
following the general pattern taken in Metalava which is to stop
using extension and override to implement model specific behavior
within the core classes (e.g. `Item`, `Codebase`, `TypeItem`). Instead
the core classes delegate to targeted interfaces (e.g. `DefaultValue`)
in which the model specific behavior can be provided. That allows the
core classes to be shared and the target interfaces clarifies the
boundaries between model independent code and model specific code.

The immediate motivation for doing this is the snapshot code needs to
be able to resolve a class that it has not yet seen. Rather than create
a `SnapshotCodebase` this work is being done to make `DefaultCodebase`
delegate that work to a `CodebaseAssembler` and then the snapshot code
can provide one of those, possibly by just having the
`CodebaseSnapshotTaker` implement it.

Bug: 353737744
Test: ./gradlew
Change-Id: I80326f86cf7c3b8f2bc1fd3e37177950648ed63e