Minor refactoring of DefaultPropertyItem for snapshot
Adds a `documentationFactory` parameter and renames
`apiVariantSelectorsFactory` to `variantSelectorsFactory` to be
consistent with other uses.

Although neither the text of turbine model needs to pass a
`documentationFactory` to the constructor (the former because it does
not track documentation and the latter because it does not handle
properties) the snapshot code does need to handle taking a snapshot of
a `PsiPropertyItem` which does track documentation so it makes sense to
do it correctly in the first place.

Bug: 353737744
Test: ./gradlew
Change-Id: Id1dce8cd77f5d0e9957c215984a6cd2ba8034646