Simplify createTypeParameterItemsAndFactory(...)
Previously, `createTypeParameterItemsAndFactory(...)` had to handle
different `TypeParameterItem` implementations so delegated the work of
setting the type bounds to the `boundsSetter` lambda. However, now that
all the `TypeParameterItem` implementations extend
`DefaultTypeParameterItem` that is no longer necessary.

This change moves the setting of the `DefaultTypeParameterItem.bounds`
into the method, changes the parameters from `boundsSetter` to
`boundsGetter` and removes the unnecessary `I` type parameter of the
method.

Bug: 352480646
Test: ./gradlew
Change-Id: Id5a1a626f07c70eb5b779c2517ceb3e926453ecc