Stop relying on normalizeInputName() in findAnnotation(String)
Metalava consumes sources that often use different annotations for the
same purposes, e.g. there are lots of duplicated annotations in the
`android.annotation` and `androidx.annotation` packages. To simplify
the processing it normalizes annotation names that it sees in the input
to use a single name internally.

Previously, (for legacy reasons) `ModifierList.findAnnotation(String)`
would assume that the `String` was the name of an annotation in the
input, before normalization and normalize it before searching. However,
that is not true. The name has either already been normalized, or is
a fixed name in the source, or on the command line. The fixed name can
simply be the normalized name. If it is necessary to support allowing
unnormalized names on the command line then it can be responsible for
normalizing them.

This change does not remove the normalization but instead treats it as
an error if the normalized name is not the same as the supplied name.
That found one place in the source code where the wrong name was used
so this change fixes that.

The consistency check is left in so that when this change is tested in
Android and AndroidX any issues will be reported directly making them
easier to fix than if it resulted in some change to a Metalava output
file. A follow-up change will remove it along with the `codebase`
property.

Bug: 356548977
Test: ./gradlew
Change-Id: Ic20ed9d57f9b2138aa6531a88067ca455f3045a5