Make DefaultModifierList.annotations not mutable
Previously, it was a `MutableList<AnnotationItem>?` so could be `null`
or a `MutableList<AnnotationItem>`. The use of `null` was a performance
optimization that unfortunately, caused issues because an empty
`DefaultModifierList` could either have an empty mutable list or a
`null` list. Callers often had to jump through hoops to produce a
`MutableList<AnnotationItem>` even if they already had a
`List<AnnotationItem>`

This changes `annotations` to be `List<AnnotationItem>`, `null` was
replaced with `emptyList()`. It fixes up any compilation issues caused
by the change but leaves any additional clean up to follow-up changes.

Bug: 356548977
Test: ./gradlew
Change-Id: I7555dbd0e2106f04907f3af92dce119a72a445ad