Use PackageTracker.createInitialPackages(...) in PsiBasedCodebase
Previously, `PsiBasedCodebase` would create the initial packages, i.e.
the ones with `package-info.java` or `package.html` files in the source
files listed on the command line, in `finishInitialization(...)`.

This change replaces that with `createInitialPackages(...)`. It also
removes the check to make sure that the root package is created because
`createInitialPackages(...)` always creates it.

Bug: 352480646
Test: ./gradlew
Change-Id: I161f5a6b1d426ba37cd7cf3e00c5dbe36fc01d7d