Refine package-info.java checks in PsiBasedCodebase
Previously, when finding the `package-info.java` files it assumed that
any `.java` file without any classes was a `package-info.java` file but
it could just have been an empty class file. Similarly, the
`getOptionalPackageDocFromPackageInfoFile(...)` method did not check to
see if the file contained a valid `PsiPackage`.

This change fixes both issues.

Bug: 352480646
Test: ./gradlew
Change-Id: If4746fc9b94a15d067689095471ef539eafdbae0